dbopsagent/app/dbopsagent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dbopsagent"
version = "0.1.0"
description = "AgentCore Runtime Application using Strands SDK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aws-opentelemetry-distro",
    "bedrock-agentcore >= 1.0.3",
    "botocore[crt] >= 1.35.0",
    "mcp >= 1.19.0",
    "pymssql>=2.2.0",
    "pyyaml>=6.0",
    "strands-agents >= 1.13.0",
    
]

[tool.hatch.build.targets.wheel]
packages = ["."]
